refactor(om): split state manager support types [skip notes]
Move DefaultStateManager and WriteFailureKind into their own files without changing state manager behavior.
