+++ Egil.Orleans.Messaging/src/Egil.Orleans.Messaging/State/WriteFailureKind.cs
@@ -1,0 +1,17 @@
+namespace Egil.Orleans.Messaging.State;
+
+/// <summary>
+/// Classifies a storage write failure for recovery decisions.
+/// </summary>
+public enum WriteFailureKind
+{
+    /// <summary>
+    /// Unknown outcome. The write may have persisted. Run read-back recovery.
+    /// </summary>
+    UnknownOutcome,
+
+    /// <summary>
+    /// The write definitely did not persist. Skip read-back and rethrow.
+    /// </summary>
+    DidNotPersist
+}
